eee-control: Remove audio hotkey handling patch
Previously audio hotkeys on the EEE did not emit XF86Audio* keypresses
and audio hotkey functions were handled in eee-control with pyalsaaudio.
Now the keypresses are emitted correctly, and this can lead to an
embarassing scenario, which I experienced firsthand: Assume audio is
muted, and an application — such as mplayer — that grabs audio hotkey
keypresses is started. Now pressing the mute button in attempt to unmute
the sound results in eee-control unmuting the volume AND mplayer muting
the audio. This is clearly stupid, and so the pyalsaaudio patch is
removed. Please map the audio hotkeys in your window manager to get the
same functionality.